fix: fix fronted end monitor sdk error and type error
