+++ sdk/index.ts
@@ -1,9 +1,7 @@
-/* eslint-disable */
 import { createBlankScreenMonitor } from './createBlankScreenMonitor'
 import { createJsErrorMonitor } from './createJsErrorMonitor'
 import { createPerfMonitor } from './createPerfMonitor'
 import { createResourceErrorMonitor } from './createResourceErrorMonitor'
-import { createXhrMonitor } from './createXhrMonitor'
 import { createPromiseErrorMonitor } from './createPromiseErrorMonitor'
 
 export function createAllMonitor(url?: string) {
@@ -16,7 +14,6 @@ export function createAllMonitor(url?: string) {
     createJsErrorMonitor(url).start()
     createPerfMonitor(url).start()
     createResourceErrorMonitor(url).start()
-    // createXhrMonitor(url).start()
     createPromiseErrorMonitor(url).start()
   }
   return { name, start }
